+++ actix-test/src/lib.rs
@@ -64,7 +64,7 @@ pub use actix_web::test::{
 ///     Ok(HttpResponse::Ok())
 /// }
 ///
-/// #[actix_rt::test]
+/// #[actix_web::test]
 /// async fn test_example() {
 ///     let srv = actix_test::start(||
 ///         App::new().service(my_handler)
@@ -104,7 +104,7 @@ where
 ///     Ok(HttpResponse::Ok())
 /// }
 ///
-/// #[actix_rt::test]
+/// #[actix_web::test]
 /// async fn test_example() {
 ///     let srv = actix_test::start_with(actix_test::config().h1(), ||
 ///         App::new().service(my_handler)