+++ actix-web-codegen/src/route.rs
@@ -220,7 +220,7 @@ fn guess_resource_type(typ: &syn::Type) -> ResourceType {
 impl Route {
     pub fn new(
         args: AttributeArgs,
-        input: TokenStream,
+        ast: syn::ItemFn,
         method: Option<MethodType>,
     ) -> syn::Result<Self> {
         if args.is_empty() {
@@ -234,14 +234,11 @@ impl Route {
                 ),
             ));
         }
-        let ast: syn::ItemFn = syn::parse(input)?;
+
         let name = ast.sig.ident.clone();
 
-        // Try and pull out the doc comments so that we can reapply them to the
-        // generated struct.
-        //
-        // Note that multi line doc comments are converted to multiple doc
-        // attributes.
+        // Try and pull out the doc comments so that we can reapply them to the generated struct.
+        // Note that multi line doc comments are converted to multiple doc attributes.
         let doc_attributes = ast
             .attrs
             .iter()
@@ -349,8 +346,28 @@ pub(crate) fn with_method(
     input: TokenStream,
 ) -> TokenStream {
     let args = parse_macro_input!(args as syn::AttributeArgs);
-    match Route::new(args, input, method) {
+
+    let ast = match syn::parse::<syn::ItemFn>(input.clone()) {
+        Ok(ast) => ast,
+        // on parse error, make IDEs happy; see fn docs
+        Err(err) => return input_and_compile_error(input, err),
+    };
+
+    match Route::new(args, ast, method) {
         Ok(route) => route.into_token_stream().into(),
-        Err(err) => err.to_compile_error().into(),
+        // on macro related error, make IDEs happy; see fn docs
+        Err(err) => input_and_compile_error(input, err),
     }
+}
+
+/// Converts the error to a token stream and appends it to the original input.
+///
+/// Returning the original input in addition to the error is good for IDEs which can gracefully
+/// recover and show more precise errors within the macro body.
+///
+/// See <https://github.com/rust-analyzer/rust-analyzer/issues/10468> for more info.
+fn input_and_compile_error(mut item: TokenStream, err: syn::Error) -> TokenStream {
+    let compile_err = TokenStream::from(err.to_compile_error());
+    item.extend(compile_err);
+    item
 }